중복 이미지 허용 및 주석 제거

+++ src/main/java/com/myboot/facilities/controller/FacilitiesControllerImpl.java
@@ -93,9 +93,11 @@ public class FacilitiesControllerImpl implements FacilitiesController {
 	    				FileUtils.copyFile(srcFile, destFile,true);
 	    			} else {
 	    				srcFile = new File(path+"\\"+"temp"+"\\"+imageFileName);
+	    				File src2File = new File(path+"\\"+f_Num+"\\"+imageFileName);
 	    				File destDir = new File(path+"\\"+f_Num);
-//						destDir.mkdirs();
-	    				FileUtils.moveFileToDirectory(srcFile, destDir,true);
+	    				if(!src2File.exists()) {
+	    					FileUtils.moveFileToDirectory(srcFile, destDir,true);
+	    				}
 	    			}
 	    		}
 	    	}	    	
@@ -117,15 +119,10 @@ public class FacilitiesControllerImpl implements FacilitiesController {
 	    	return resEnt;
 		}	    
 	    
-
-	
-//	  	다중이미지 업로드
 		private Map<String,String> upload(MultipartHttpServletRequest multipartRequest) 
 		throws Exception {
-//			List<String> fileList= new ArrayList<String>();
 			Map<String,String>fileMap = new HashMap<String,String>();
 			Iterator<String> fileNames = multipartRequest.getFileNames();
-//			============================================ realPath 받아오기
 			String realPath = multipartRequest.getSession().getServletContext().getRealPath("");
 			String path = realPath+"resources\\facilities\\facilities_image";
 
@@ -133,10 +130,9 @@ public class FacilitiesControllerImpl implements FacilitiesController {
 				String fileName = fileNames.next();
 				MultipartFile mFile = multipartRequest.getFile(fileName);
 				String originalFileName = mFile.getOriginalFilename();
-//				fileList.add(originalFileName);
 				fileMap.put(fileName, originalFileName);
-				File file = new File(path +"\\"+"temp"+"\\" + fileName);
-				if(mFile.getSize()!=0) { // File Null Check
+				File file = new File(path +"\\"+"temp"+"\\" + originalFileName);
+				if(mFile.getSize()!=0) { 
 					if(!file.exists()) {
 						file.getParentFile().mkdirs();
 						mFile.transferTo(new File(path +"\\"+"temp"+ "\\"+originalFileName));
